infra: hot-reload on dev enviroment
