+++ docker-compose.yml
@@ -26,6 +26,7 @@ services:
       - .env
     volumes:
       - .:/app
+      - /app/node_modules
     command: npm run dev
 
 volumes:
