+++ package.json
@@ -26,5 +26,11 @@
   },
   "prisma": {
     "seed": "node prisma/seed.js"
+  },
+  "nodemonConfig": {
+    "watch": ["src"],
+    "ext": "js,json",
+    "ignore": ["node_modules"],
+    "legacyWatch": true
   }
 }
